requirements.txt
# Core data science libraries
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Jupyter and notebook tools
jupyter>=1.0.0
jupyterlab>=4.0.0
ipykernel>=6.25.0

# Data visualization and analysis
altair>=5.0.0
bokeh>=3.2.0
dash>=2.10.0

# Database and data handling
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pymongo>=4.4.0

# Statistical analysis
statsmodels>=0.14.0
pingouin>=0.5.0

# Machine learning extensions
xgboost>=1.7.0
lightgbm>=4.0.0
catboost>=1.2.0

# Geospatial data (useful for migration tracking)
geopandas>=0.13.0
folium>=0.14.0
shapely>=2.0.0

# Web scraping and APIs
requests>=2.31.0
beautifulsoup4>=4.12.0

# Configuration and utilities
python-dotenv>=1.0.0
pyyaml>=6.0
click>=8.1.0
tqdm>=4.65.0

# Development tools
pytest>=7.4.0
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0
